+++ RELEASE_NOTES.md
@@ -1,6 +1,18 @@
 Release Notes
 =============
 
+v1.0.0-alpha.X
+--------------
+
+### Breaking changes
+
+- Split `Specification` class into `SpecificationBase` and `TraitsData`
+  This properly defines the separation of the generic data container
+  from the strongly typed views used to get/set well known traits and
+  their properties.
+  [#348](https://github.com/OpenAssetIO/OpenAssetIO/issues/348)
+
+
 v1.0.0-alpha.1
 --------------